+++ TAO/TAO_IDL/ast/ast_decl.cpp
@@ -84,84 +84,6 @@ trademarks or registered trademarks of Sun Microsystems, Inc.
 // FUZZ: disable check_for_streams_include
 #include "ace/streams.h"
 
-AST_Annotation_Appls::AST_Annotation_Appls ()
-{
-}
-
-AST_Annotation_Appls::AST_Annotation_Appls (const AST_Annotation_Appls& other)
-{
-  *this = other;
-}
-
-AST_Annotation_Appls::~AST_Annotation_Appls ()
-{
-}
-
-AST_Annotation_Appls &
-AST_Annotation_Appls::operator= (const AST_Annotation_Appls& other)
-{
-  vector_.clear ();
-  add (other);
-  return *this;
-}
-
-void
-AST_Annotation_Appls::add (AST_Annotation_Appl *appl)
-{
-  vector_.push_back (AST_Annotation_Appl_Ptr (appl));
-}
-
-void
-AST_Annotation_Appls::add (const AST_Annotation_Appls &other)
-{
-  for (const_iterator i = other.begin (); i != other.end (); ++i)
-    {
-      vector_.push_back (*i);
-    }
-}
-
-bool
-AST_Annotation_Appls::empty () const
-{
-  return !vector_.size ();
-}
-
-size_t
-AST_Annotation_Appls::size () const
-{
-  return vector_.size ();
-}
-
-AST_Annotation_Appls::iterator
-AST_Annotation_Appls::begin ()
-{
-  return vector_.begin ();
-}
-
-AST_Annotation_Appls::iterator
-AST_Annotation_Appls::end ()
-{
-  return vector_.end ();
-}
-
-AST_Annotation_Appls::const_iterator
-AST_Annotation_Appls::begin () const
-{
-  return vector_.begin ();
-}
-
-AST_Annotation_Appls::const_iterator
-AST_Annotation_Appls::end () const
-{
-  return vector_.end ();
-}
-
-AST_Annotation_Appl *
-AST_Annotation_Appls::operator[] (size_t index)
-{
-  return vector_[index].get ();
-}
-
 COMMON_Base::COMMON_Base (bool local,
                           bool abstract)
   : is_local_ (local),
